Section 8: Until 96th Video Lecture

+++ flask-demo-app/resources/item.py
@@ -1,11 +1,11 @@
 import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 import models
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from schemas import ItemSchema, ItemUpdateSchema
+from flask_jwt_extended import jwt_required, get_jwt
 
 blp = Blueprint("items", __name__, description="Operations on items")
 
@@ -17,11 +17,16 @@ class Item(MethodView):
         item = models.ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        print(jwt)
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin priviledge required")
         item = models.ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
-        return {"message": f"Item {item.name} Deleted"}        
+        return {"message": f"Item {item.name} Deleted", "token": f"{jwt}"}        
 
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemUpdateSchema)
@@ -42,10 +47,12 @@ class Item(MethodView):
 @blp.route("/item")
 class ItemList(MethodView):
 
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return models.ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
